fixing title input behavior on new/existing button usage
Editing new-review-title should be inactive if add to existing is
selected, and the review list/refresh button should be active. The title
entry box should be active and the other two should be inactive if
create new review is selected.  This sorts out the boolean values
controlling that in each case.