fixing git/svn review creation/update
There was a subtle bug affecting how we described the changelist(s)
being submitted to create (or update) a review from, that was preventing
the Collaborator web ui from properly displaying the reviews
(changelists from different clients were not properly coalescing).  This
fixes that; sadly that means pulling in a bunch of external SCM
libraries because there is no obvious way to get the required repo
metadata out of the IntelliJ openapi.

Bump plugin version number to 0.7.  I think this is now stable enough to
release as a beta.